feat(core): 实现商品变更事件处理机制
- 新增 ProductChangedEvent 和 BannerChangedEvent 类用于商品和轮播图变更事件
- 在 ProductServiceImpl 和 BannerServiceImpl 中添加事件处理逻辑
- 实现 syncProductToEs、deleteProductFromEs 和 refreshCache 方法
- 更新相关控制器和 Redis 服务实现

+++ mall-banner/src/main/java/com/jiang/mall/controller/BannerAdminController.java
@@ -177,7 +177,7 @@ public class BannerAdminController {
      * @return 删除结果或错误信息
      */
     @GetMapping("/delete")
-    public ResponseResult<Object> deleteBanner(@RequestParam("id") Integer id,
+    public ResponseResult<Object> deleteBanner(@RequestParam("id") Long id,
                                                 HttpSession session) {
         if (id==null||id<=0){
             return ResponseResult.failResult("参数错误");
